+++ nls/root/strings.js
@@ -3,6 +3,8 @@
 define({
     ADD_ENDLINE_TO_THE_END_OF_FILE:     "Add endline at the end of file",
     ADD_TO_GITIGNORE:                   "Add to .gitignore",
+    AMEND_COMMIT:                       "Amend last commit",
+    AMEND_COMMIT_FORBIDDEN:             "Cannot amend commit when there are no unpushed commits",
     REMOVE_FROM_GITIGNORE:              "Remove from .gitignore",
     BRACKETS_GIT_ERROR:                 "Brackets Git encountered an error\u2026",
     BRANCH_NAME:                        "Branch name",